Find the First Connection for a Given Sentence Finds the first connection for a given sentence in a list of sentences.

Find the First Connection for a Given Sentence Finds the first connection for a given sentence in a list of sentences.

Usage

connect_with(sentences, i)

Arguments

sentences

A list of sentences (character vectors).

i

An index of the target sentence

Value

A list representing the connection (with elements sentence_i, sentence_j, word_i, and word_j), or only sentence_i when no connection found.

Examples

sentences <- list(
  c("there", "is", "a", "quick", "brown", "fox"),
  c("the", "fox", "chase", "a", "lazy", "dog"),
  c("dog", "jump", "over", "the", "moon")
)
connect_with(sentences, 1)
#> $sentence_i
#> [1] 1
#> 
#> $word_i
#> [1] 1
#> 
connect_with(sentences, 2)
#> $sentence_i
#> [1] 2
#> 
#> $sentence_j
#> [1] 1
#> 
#> $word_i
#> [1] 2
#> 
#> $word_j
#> [1] 6
#> 
connect_with(sentences, 3)
#> $sentence_i
#> [1] 3
#> 
#> $sentence_j
#> [1] 2
#> 
#> $word_i
#> [1] 1
#> 
#> $word_j
#> [1] 6
#>