Find the First Common Word in Two Sentences

Finds the first common word in two sentences and returns its indices in each sentence.

Usage

search_common_word(sentence_i, sentence_j)

Arguments

sentence_i, sentence_j

A character vector of sentence.

Value

A list with two elements or NULL when no common word was found:

common_i: The index of the common word in the sentence of i.

common_j: The index of the common word in the sentence of j.

Examples

sentence1 <- c("the", "quick", "brown", "fox")
sentence2 <- c("the", "lazy", "dog")
search_common_word(sentence1, sentence2)
#> $common_i
#> [1] 1
#> 
#> $common_j
#> [1] 1
#>